feat: dataProvider interface
Adds an API layer between the frontend and the data sources defined as
the `DataProvider` interface. This first version only applies to the
Voter App and supports only a `StrapiDataProvider` and a partial
implementation of a `LocalDataProvider`.

The feedback implementation of `LocalDataProvider` is accessed via an
`/api/feedack` route, which writes the feedback as a JSON file in
`/data/feedback`. This is used as a fallback when the `DataProvider`
implementation used does not support receiving feedback.

Additionally:

- Adds a `dataProvider` setting that defines the
`DataProvider` to use and specifies whether the Candidate App is
supported.
- Refactors the `StrapiDataProvider` module to multiple files.
- Refactors the `color` utilities and moves out of the `api` folder to
  `utils`.
- Excludes API routes from locale-based redirection in
  `hooks.server.ts`.
- Adds the `/data` folder to gitignore.
