+++ frontend/src/lib/api/dataProvider/index.ts
@@ -1,0 +1,1 @@
+export * from './dataProvider';
