+++ frontend/src/lib/api/dataProvider/strapi/index.ts
@@ -1,0 +1,2 @@
+export * from './strapiDataProvider';
+export * from './strapiDataProvider.type';
