+++ frontend/src/lib/api/dataProvider/strapi/strapiDataProvider.type.ts
@@ -261,6 +261,7 @@ export interface StrapiQuestionCategoryData {
     questions: {
       data: StrapiQuestionData[];
     };
+    customData: JSONData;
   };
 }
