+++ frontend/src/lib/api/sendFeedback.ts
@@ -1,0 +1,19 @@
+import {browser} from '$app/environment';
+import type {FeedbackData} from './dataProvider/dataProvider';
+import {dataProvider} from './getData';
+
+export async function sendFeedback(
+  rating?: FeedbackData['rating'],
+  description?: FeedbackData['description']
+): Promise<Response | undefined> {
+  if (!browser) return Promise.resolve(undefined);
+  const {setFeedback} = await dataProvider;
+  const data: FeedbackData = {
+    rating,
+    description,
+    date: new Date(),
+    url: window?.location?.href,
+    userAgent: navigator?.userAgent
+  };
+  return setFeedback(data);
+}
