--- frontend/src/lib/utils/color/index.ts
@@ -1,1 +1,0 @@
-export * from './color';
