+++ frontend/src/lib/utils/color/rgb.ts
@@ -1,0 +1,3 @@
+/** An RGB color array. The numbers' domain is [0, 255] */
+
+export type RGB = [number, number, number];
