+++ frontend/src/routes/[[lang=locale]]/(voters)/questions/+layout.server.ts
@@ -1,8 +1,9 @@
 import type {LayoutServerLoad} from './$types';
-import {getOpinionQuestions} from '$lib/api/getData';
+import {dataProvider} from '$lib/api/getData';
 
 export const load = (async ({parent}) => {
   const locale = (await parent()).i18n.currentLocale;
+  const {getOpinionQuestions} = await dataProvider;
   return {
     opinionQuestions: getOpinionQuestions({locale})
   };
