+++ frontend/src/routes/[[lang=locale]]/(voters)/results/+layout.server.ts
@@ -1,13 +1,10 @@
-import {
-  getInfoQuestions,
-  getNominatedCandidates,
-  getNominatingParties,
-  getOpinionQuestions
-} from '$lib/api/getData';
+import {dataProvider} from '$lib/api/getData';
 import type {LayoutServerLoad} from './$types';
 
 export const load = (async ({parent}) => {
   const locale = (await parent()).i18n.currentLocale;
+  const {getInfoQuestions, getNominatedCandidates, getNominatingParties, getOpinionQuestions} =
+    await dataProvider;
   return {
     candidates: getNominatedCandidates({loadAnswers: true, locale}),
     parties: getNominatingParties({loadAnswers: true, locale}),
